+++ src/vue/pages/bridge/tabs/bridge-tokens/BridgeConfirmation.vue
@@ -31,6 +31,15 @@
         <p class="bridge-confirmation__chain-status-msg">
           {{ chainStatusMsg }}
         </p>
+
+        <app-button
+          v-if="!isToChainActive"
+          class="bridge-confirmation__connect-chain-btn"
+          scheme="primary"
+          :text="$t('bridge-page.bridge-confirmation.connect-chain-btn')"
+          :disabled="isConnectChainBtnDisabled"
+          @click="connectEthereumChain"
+        />
       </div>
 
       <app-button
@@ -58,6 +67,7 @@ import { Bus } from '@/js/helpers/event-bus'
 import { ChainRecord } from '@/js/records/chain.record'
 import { TokenRecord } from '@/js/records/token.record'
 import { Erc721TokenRecord } from '@/js/records/erc721-token.record'
+import { switchOrAddEthereumChain } from '@/js/helpers/metamask-helper'
 
 const STEPS = {
   depositOrWithdrawFirstStep: 1,
@@ -113,6 +123,7 @@ export default {
       txHash: '',
       currentStep: STEPS.depositOrWithdrawFirstStep,
       processing: false,
+      isConnectChainBtnDisabled: false,
     })
 
     const isToChainActive = computed(() =>
@@ -311,6 +322,24 @@ export default {
       state.processing = false
     }
 
+    async function connectEthereumChain () {
+      state.isConnectChainBtnDisabled = true
+
+      try {
+        await switchOrAddEthereumChain({
+          hexId: props.toChain.hexId,
+          nativeName: props.toChain.nativeName,
+          rpcUrl: props.toChain.rpcUrl,
+          blockExplorerUrl: props.toChain.blockExplorerUrl,
+          nativeSymbol: props.toChain.nativeSymbol,
+          nativeDecimals: props.toChain.nativeDecimals,
+        })
+      } catch (e) {
+        ErrorHandler.processWithoutFeedback(e)
+      }
+      state.isConnectChainBtnDisabled = false
+    }
+
     return {
       ...toRefs(state),
       chainStatusMsg,
@@ -320,6 +349,7 @@ export default {
       depositOrWithdrawSecondStep,
       isButtonDisabled,
       buttonTranslation,
+      connectEthereumChain,
       STEPS,
     }
   },
@@ -362,6 +392,10 @@ export default {
   font-size: 1.4rem;
   max-width: 25rem;
   text-align: center;
+}
+
+.bridge-confirmation__connect-chain-btn {
+  margin-top: 2rem;
 }
 
 .bridge-confirmation__mint-btn {